Fix a problem Mike found with the recent check-in.

+++ mag/harp/libs/matlab/mfile-plain/track_hmi_loop.m
@@ -665,7 +665,7 @@ for f = 1:nf,
     % 11/2012: put in smarter update to not overwrite good info with zeros
     chip_now = chip1(inx1,inx2);
     % chip1(inx1,inx2) = full(rois1(r).chip); % (original - bug)
-    chip1(inx1,inx2) = max(chip_now, full(rois1(r).chip)); % (new)
+    chip1(inx1,inx2) = max(chip_now, uint8(full(rois1(r).chip))); % (new)
   end;
   sp(f).chip = chip1; % plug it in
 end;
